cart

+++ spring-backend/src/main/java/com/springboot/laptop/model/CartDetails.java
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
@@ -17,7 +16,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class CartDetails extends  BaseEntity{
 
-    private Integer quantity;
+    private Long quantity;
 
     private LocalDateTime addDate;
 
@@ -33,6 +32,5 @@ public class CartDetails extends  BaseEntity{
     @ManyToOne
     @JoinColumn(name="product_id", referencedColumnName = "id")
     private ProductEntity product;
-
 
 }
